refact: move binding structures to the module package
