+++ std/injector/src/test/java/com/jeanbarrossilva/orca/std/injector/InjectorTests.kt
@@ -2,9 +2,9 @@ package com.jeanbarrossilva.orca.std.injector
 
 import assertk.assertThat
 import assertk.assertions.isEqualTo
-import com.jeanbarrossilva.orca.std.injector.binding.boundTo
 import com.jeanbarrossilva.orca.std.injector.module.Inject
 import com.jeanbarrossilva.orca.std.injector.module.Module
+import com.jeanbarrossilva.orca.std.injector.module.binding.boundTo
 import com.jeanbarrossilva.orca.std.injector.test.InjectorTestRule
 import kotlin.test.Test
 import org.junit.Rule
